Clean code: Move Raptor test data helpers into the '_data' package, and make sub packages 'stoparrival' and 'transit'.
