+++ src/test/java/org/opentripplanner/transit/raptor/_data/stoparrival/Egress.java
@@ -1,5 +1,6 @@
-package org.opentripplanner.transit.raptor._shared;
+package org.opentripplanner.transit.raptor._data.stoparrival;
 
+import org.opentripplanner.transit.raptor._data.transit.TestTripSchedule;
 import org.opentripplanner.transit.raptor.api.view.ArrivalView;
 import org.opentripplanner.transit.raptor.util.TimeUtils;
