+++ src/test/java/org/opentripplanner/transit/raptor/_data/transit/TestTripSchedule.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.transit.raptor._shared;
+package org.opentripplanner.transit.raptor._data.transit;
 
 import org.opentripplanner.model.base.ToStringBuilder;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripPattern;
